backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pgvector
pydantic
pydantic-settings
alembic
python-dotenv
google-genai
supabase
